Update fetch queries for page types in webapp/fullstack examples (#28)
## What is the goal of this PR?

Simplify the portion of the `fetch` queries in the webapp and fullstack
examples where they retrieve the type (person/organization/group) of a
given page.

## What are the changes implemented in this PR?

We match the label of the type, and then return the entire type object
without constructing intermediary string representations. We update the
frontend to use this wrapped object instead of the string directly.